requirements.txt
yt-dlp==2025.1.26
faster-whisper==1.1.0

google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0